requirements/dev.txt
-r common.txt
pytest==6.2.5
pytest-httpserver==1.0.4
pydantic-factories==1.2.3
pre-commit==2.17.0
pytest-cases==3.6.8
pytest-cov==3.0.0
wheel==0.38.0
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
